Move plugin type to separate file
