+++ tools/readme_linter/main.go
@@ -4,40 +4,11 @@ import (
 	"bufio"
 	"bytes"
 	"os"
-	"strings"
 
 	"github.com/yuin/goldmark"
 	"github.com/yuin/goldmark/ast"
 	"github.com/yuin/goldmark/text"
 )
-
-type plugin int
-
-const (
-	pluginNone plugin = iota
-	pluginInput
-	pluginOutput
-	pluginProcessor
-	pluginAggregator
-	pluginParser
-)
-
-func guessPluginType(filename string) plugin {
-	switch {
-	case strings.Contains(filename, "plugins/inputs/"):
-		return pluginInput
-	case strings.Contains(filename, "plugins/outputs/"):
-		return pluginOutput
-	case strings.Contains(filename, "plugins/processors/"):
-		return pluginProcessor
-	case strings.Contains(filename, "plugins/aggregators/"):
-		return pluginAggregator
-	case strings.Contains(filename, "plugins/parsers/"):
-		return pluginParser
-	default:
-		return pluginNone
-	}
-}
 
 func main() {
 	var err error
